- Sync WIDL and ole32 with Wine-1.1.23, w/o syncing rpcrt4. Credits go to Christoph for helping out with this sync. Tested with rpcrt4_winetests and MS Office 2003 setup.
svn path=/trunk/; revision=42480

+++ reactos/tools/widl/hash.c
@@ -23,6 +23,7 @@
 
 #include <host/nls.h>
 
+#include "widltypes.h"
 #include "hash.h"
 
 static const unsigned char Lookup_16[128 * 3] = {
@@ -500,7 +501,7 @@ static const unsigned char Lookup_224[128 * 3] = {
  *  skind and lcid, while the low word is based on a repeated string
  *  hash of skind/str.
  */
-unsigned long lhash_val_of_name_sys( syskind_t skind, LCID lcid, LPCSTR lpStr)
+unsigned int lhash_val_of_name_sys( syskind_t skind, LCID lcid, LPCSTR lpStr)
 {
   ULONG nOffset, nMask = skind == SYS_MAC ? 1 : 0;
   ULONG nHiWord, nLoWord = 0x0deadbee;
